+++ .minecraft/server_distribution/cleanup_stale_jars.ps1
@@ -46,6 +46,7 @@ $customJars = @(
     'iridescent_origins-1.0.0.jar',
     'iridescent_biomes-1.0.0.jar',
     'iridescent_modular_spells-0.2.0.jar',
+    'iridescent_durability_clamp-0.1.0.jar',
     'mek_walkable_cables-1.0.1.jar',
     'offlineskins-1.20.1-v1.jar',
     'zeta_racefix-1.0.0.jar',